Template class to own file
